requirements.txt
fastapi
uvicorn
sqlmodel
pyodbc
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
fastapi[all]
python-multipart
pydantic-settings